Streamline and document types
